log rejected operation in more socket tests
